requirements.txt
Django
python-dotenv
supabase
psycopg2
psycopg2-binary
djangorestframework
djangorestframework-simplejwt
django-environ